+++ walk2.c
@@ -61,7 +61,7 @@ static inline int getCell(PKD pkd,int iCache,int iCell,int id,float *pcOpen,KDN 
 
 
 #ifdef USE_SIMD_OPEN
-void iOpenOutcomeSIMD(PKD pkd,KDN *k,CL cl,CLTILE tile,float dThetaMin,int SPHoptions);
+void iOpenOutcomeSIMD(PKD pkd,KDN *k,CL cl,CLTILE tile,float dThetaMin,uint64_t SPHoptions);
 #endif
 #if 1
 /*
@@ -71,7 +71,7 @@ void iOpenOutcomeSIMD(PKD pkd,KDN *k,CL cl,CLTILE tile,float dThetaMin,int SPHop
 **
 ** This version has been changed by adding the ability to open buckets.
 */
-static void iOpenOutcomeCL(PKD pkd,KDN *k,CL cl,CLTILE tile,float dThetaMin,int SPHoptions) {
+static void iOpenOutcomeCL(PKD pkd,KDN *k,CL cl,CLTILE tile,float dThetaMin,uint64_t SPHoptions) {
     const int walk_min_multipole = 3;
     float dx,dy,dz,mink2,d2,d2Open,xc,yc,zc,fourh2,minbnd2,kOpen,cOpen,diCrit;
     int i;
@@ -172,7 +172,7 @@ static void addChild(PKD pkd, int iCache, CL cl, int iChild, int id, float *fOff
 static int processCheckList(PKD pkd, SMX smx, SMF smf, int iRoot, int iRoot2, 
     struct pkdKickParameters *kick,struct pkdLightconeParameters *lc,struct pkdTimestepParameters *ts,
     double dTime,int bEwald,
-    double dThetaMin, double *pdFlop, double *pdPartSum,double *pdCellSum,int SPHoptions) {
+    double dThetaMin, double *pdFlop, double *pdPartSum,double *pdCellSum,uint64_t SPHoptions) {
     KDN *k,*c,*kFind;
     int id,idUpper,iCell,iSib,iLower,iUpper,iCheckCell,iCheckLower,iCellDescend;
     PARTICLE *p;
@@ -807,7 +807,7 @@ int pkdGravWalkHop(PKD pkd,double dTime,int nGroup, double dThetaMin,double *pdF
 int pkdGravWalk(PKD pkd,struct pkdKickParameters *kick,struct pkdLightconeParameters *lc,struct pkdTimestepParameters *ts,
     double dTime,int nReps,int bEwald,int nGroup,
     int iLocalRoot1, int iLocalRoot2,int iVARoot,
-    double dThetaMin,double *pdFlop,double *pdPartSum,double *pdCellSum,int SPHoptions) {
+    double dThetaMin,double *pdFlop,double *pdPartSum,double *pdCellSum,uint64_t SPHoptions) {
     int id;
     float fOffset[3];
     int ix,iy,iz,bRep;
